Add remote URL to fetch exceptions
When an exception occurs during fetching a remote URL, it's often useful
to see the remote URL, instead of having to guess. This catches the
original Exception (usually from Requests) and re-raises with the remote
URL as additional info in the message.

Maybe we have to expand the try block to catch more exceptions.